requirements.txt
Pillow==5.2.0
numpy==1.14.5